nuevos Cambios

+++ src/main/java/com/redsocial/servicio/ConcursoServicio.java
@@ -7,7 +7,7 @@ import com.redsocial.entidad.Concurso;
 
 public interface ConcursoServicio  {
 
-	public abstract void eliminaConcurso(int id) ;
+	public abstract void eliminaConcurso(long id) ;
 	public abstract Concurso insertaActualizaConcurso(Concurso  obj) ;
 	public abstract List<Concurso> listaConcurso() ;
 	public abstract List<Concurso> listaConcursoPorNombre(String filtro) ;
